country codes adgusted
combo box options
 regulated by regular expressions

+++ Hotel-Management/Hotel-Management/Form_ClientInfo.cs
@@ -9,6 +9,8 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Configuration;
+using System.Text.RegularExpressions;
+
 namespace Hotel_Management
 {
     public partial class Form_ClientInfo : Form
@@ -32,17 +34,57 @@ namespace Hotel_Management
             con.Close();
             
         }
+        private bool validateinput()
+        {
+            Regex regexphoneETH = new Regex("^[+][0-9]{11}$");
+            Regex regexphoneKen = new Regex("^[+][0-9]{10}$");
+            Regex regexphoneGha = new Regex("^[+][0-9]{8}$");
+            Regex regexphoneEgt = new Regex("^[+][0-9]{7}$");
+            Regex regexphoneIsr = new Regex("^[+][0-9]{10}$");
+            Regex regexphoneChi = new Regex("^[+][0-9]{12}$");
+            Regex regexphoneUs = new Regex("^[+][0-9]{10}$");
+            Regex regexphoneCan = new Regex("^[+][0-9]{10}$");
+            Regex regexphoneUk = new Regex("^[+][0-9]{9}$");
+            bool valid = true;
+            if (txt_ClientPhoneNumber.Text.Equals(string.Empty) || 
+                regexphoneETH.IsMatch(txt_ClientPhoneNumber.Text)||
+                regexphoneKen.IsMatch(txt_ClientPhoneNumber.Text) ||
+                regexphoneGha.IsMatch(txt_ClientPhoneNumber.Text) ||
+                regexphoneEgt.IsMatch(txt_ClientPhoneNumber.Text) ||
+                regexphoneIsr.IsMatch(txt_ClientPhoneNumber.Text) ||
+                regexphoneChi.IsMatch(txt_ClientPhoneNumber.Text) ||
+                regexphoneUs.IsMatch(txt_ClientPhoneNumber.Text) ||
+                regexphoneCan.IsMatch(txt_ClientPhoneNumber.Text) ||
+                regexphoneUk.IsMatch(txt_ClientPhoneNumber.Text) 
+                )
+         //country digits   //kenya 10 // ghana 8  //egypt 7 //israel 10 //china 11  //us 10 //canada 10 //uk 9
+            {
+                valid = false;
+                txt_ClientPhoneNumber.Focus();
+                errorProviderforclient.SetError(txt_ClientPhoneNumber, "Invalid ENTRY, Please Enter valid phone number ");
+            }
+            if (txt_ClientID.Text.Equals(string.Empty))
+            {
+                valid = false;
+                txt_ClientID.Focus();
+                errorProviderforclient.SetError(txt_ClientID, "Invalid ENTRY, Please Enter client id ");
+            }
+            if (txt_ClientName.Text.Equals(string.Empty))
+            {
+                valid = false;
+                txt_ClientID.Focus();
+                errorProviderforclient.SetError(txt_ClientName, "Invalid ENTRY, Please Enter your name ");
+            }
+            if (comboBox1.SelectedItem==null)
+            {
+                valid = false;
+                comboBox1.Focus();
+                errorProviderforclient.SetError(comboBox1, "Invalid ENTRY, Please Enter Among the countries");
+            }
+            return valid;
+        }
         public void demo()
         {
-/*Ethiopia
-Kenya
-Ghana
-Egypt
-Israel
-China
-United States
-Canada
-United Kingdom*/
             if (comboBox1.SelectedItem.ToString().Equals("Ethiopia"))
             {
                 txt_ClientPhoneNumber.Text = "+251";
@@ -89,31 +131,37 @@ United Kingdom*/
 
         private void label_Add_Click(object sender, EventArgs e)
         {
-            SqlConnection con = new SqlConnection(constring);
-            con.Open();
-            SqlCommand Command = new SqlCommand("insert into Client values(@ClientID,@ClientName,@ClientPhone,@ClientCountry)", con);
-            Command.Parameters.AddWithValue("@ClientID", txt_ClientID.Text);
-            Command.Parameters.AddWithValue("@ClientName", txt_ClientName.Text);
-            Command.Parameters.AddWithValue("@ClientPhone", txt_ClientPhoneNumber.Text);
-            Command.Parameters.AddWithValue("@ClientCountry", comboBox1.SelectedItem.ToString());
+            if (validateinput())
+            {
+                SqlConnection con = new SqlConnection(constring);
+                con.Open();
+                SqlCommand Command = new SqlCommand("insert into Client values(@ClientID,@ClientName,@ClientPhone,@ClientCountry)", con);
+                Command.Parameters.AddWithValue("@ClientID", txt_ClientID.Text);
+                Command.Parameters.AddWithValue("@ClientName", txt_ClientName.Text);
+                Command.Parameters.AddWithValue("@ClientPhone", txt_ClientPhoneNumber.Text);
+                Command.Parameters.AddWithValue("@ClientCountry", comboBox1.SelectedItem.ToString());
 
-            Command.ExecuteNonQuery();
-            MessageBox.Show("Client Added Successfully!!!");
-            con.Close();
-            populate();
+                Command.ExecuteNonQuery();
+                MessageBox.Show("Client Added Successfully!!!");
+                con.Close();
+                populate();
+            }
         }
 
         private void label_Delete_Click(object sender, EventArgs e)
         {
-            SqlConnection con = new SqlConnection(constring);
-            con.Open();
-            SqlCommand Command = new SqlCommand("delete from Client where @ClientID=ClientID", con);
-            Command.Parameters.AddWithValue("@ClientID", txt_ClientID.Text);
+            if (validateinput())
+            {
+                SqlConnection con = new SqlConnection(constring);
+                con.Open();
+                SqlCommand Command = new SqlCommand("delete from Client where @ClientID=ClientID", con);
+                Command.Parameters.AddWithValue("@ClientID", txt_ClientID.Text);
 
-            Command.ExecuteNonQuery();
-            MessageBox.Show("Client Deleted Successfully!!!");
-            con.Close();
-            populate();
+                Command.ExecuteNonQuery();
+                MessageBox.Show("Client Deleted Successfully!!!");
+                con.Close();
+                populate();
+            }
         }
 
         private void label_Exit_Click(object sender, EventArgs e)
